python-playground/requirements.txt
selenium~=4.35.0
requests~=2.32.5
python-dotenv~=1.1.1
webdriver-manager~=4.0.2